A few fixes have been done so it works now properly
